+++ src/ts/xr.ts
@@ -17,81 +17,50 @@
 
 import "../styles/index.scss";
 
-import { Assets } from "./assets/assets";
-import { Quaternion, Vector3 } from "@babylonjs/core/Maths/math.vector";
-import { Engine } from "@babylonjs/core/Engines/engine";
-import { Scene } from "@babylonjs/core/scene";
-import { StarfieldPostProcess } from "./postProcesses/starfieldPostProcess";
+import { Vector3 } from "@babylonjs/core/Maths/math.vector";
+import "@babylonjs/core/Materials/standardMaterial";
+import "@babylonjs/core/Loading/loadingScreen";
+import "@babylonjs/core/Misc/screenshotTools";
+import "@babylonjs/core/Rendering/depthRendererSceneComponent";
 import { TransformNode } from "@babylonjs/core/Meshes/transformNode";
+import "@babylonjs/core/Meshes/thinInstanceMesh";
 import { Mandelbulb } from "./anomalies/mandelbulb/mandelbulb";
 import { MandelbulbPostProcess } from "./anomalies/mandelbulb/mandelbulbPostProcess";
+import { Scene } from "@babylonjs/core/scene";
+import { JuliaSet } from "./anomalies/julia/juliaSet";
+import { JuliaSetPostProcess } from "./anomalies/julia/juliaSetPostProcess";
+import { Color4 } from "@babylonjs/core/Maths/math.color";
 import "@babylonjs/core";
-import { FreeCamera } from "@babylonjs/core/Cameras/freeCamera";
+import { ArcRotateCamera, Engine } from "@babylonjs/core";
 
 const canvas = document.getElementById("renderer") as HTMLCanvasElement;
 canvas.width = window.innerWidth;
 canvas.height = window.innerHeight;
 
-const engine = new Engine(canvas, true);
-engine.useReverseDepthBuffer = true;
+const engine = new Engine(canvas);
 engine.displayLoadingUI();
 
 const scene = new Scene(engine);
-scene.useRightHandedSystem = true;
+scene.clearColor = new Color4(0, 0, 0, 0);
 
-const fallbackCamera = new FreeCamera("FallbackCamera", new Vector3(0, 0, 0), scene);
-fallbackCamera.attachControl(canvas, true);
-scene.enableDepthRenderer(fallbackCamera, false, true);
+const urlParams = new URLSearchParams(window.location.search);
 
-await Assets.Init(scene);
+const camera = new ArcRotateCamera("ArcRotateCamera", 0, 3.14 / 3, 5, Vector3.Zero(), scene);
+camera.attachControl(canvas, true);
+camera.lowerRadiusLimit = 0.5;
+camera.wheelPrecision *= 100;
 
-const xr = await scene.createDefaultXRExperienceAsync();
-if (!xr.baseExperience) {
-    // no xr support
-    throw new Error("No XR support");
-} else {
-    // all good, ready to go
-    console.log("XR support");
-}
-
-const webXRInput = xr.input; // if using the experience helper, otherwise, an instance of WebXRInput
-webXRInput.onControllerAddedObservable.add((xrController) => {
-    console.log("Controller added");
-    xrController.onMotionControllerInitObservable.add((motionController) => {
-        console.log("Motion controller initialized");
-
-        const mainComponent = motionController.getMainComponent();
-
-        mainComponent.onButtonStateChangedObservable.add((component) => {
-            if (component.changes.pressed) {
-                if (component.changes.pressed.current) {
-                    console.log("Pressed");
-                }
-                if (component.pressed) {
-                    console.log("Pressed");
-                }
-            }
-        });
-    });
-});
-
-const xrCamera = xr.baseExperience.camera;
-xrCamera.rigCameras.forEach((camera) => {
-    scene.enableDepthRenderer(camera, false, true);
-});
-
-const starfieldPostProcess = new StarfieldPostProcess(scene, [], [], Quaternion.Identity());
-fallbackCamera.attachPostProcess(starfieldPostProcess);
-xrCamera.attachPostProcess(starfieldPostProcess);
+const depthRenderer = scene.enableDepthRenderer(null, false, true);
 
 function createMandelbulb(): TransformNode {
     const mandelbulb = new Mandelbulb("bulb", scene, Math.random() * 10000, null);
-    mandelbulb.getTransform().setAbsolutePosition(new Vector3(0, 0, 20));
-    mandelbulb.getTransform().scalingDeterminant = 1 / 100e3;
+    mandelbulb.getTransform().scalingDeterminant = 1 / 400e3;
 
     const mandelbulbPP = new MandelbulbPostProcess(mandelbulb, scene, []);
-    fallbackCamera.attachPostProcess(mandelbulbPP);
-    xrCamera.attachPostProcess(mandelbulbPP);
+    scene.cameras.forEach(camera => camera.attachPostProcess(mandelbulbPP));
+    scene.onNewCameraAddedObservable.add((camera) => {
+        camera.attachPostProcess(mandelbulbPP);
+    });
 
     scene.onBeforeRenderObservable.add(() => {
         const deltaSeconds = engine.getDeltaTime() / 1000;
@@ -101,7 +70,46 @@ function createMandelbulb(): TransformNode {
     return mandelbulb.getTransform();
 }
 
-createMandelbulb();
+function createJulia(): TransformNode {
+    const julia = new JuliaSet("Julia", scene, Math.random() * 10000, null);
+    julia.getTransform().scalingDeterminant = 1 / 400e3;
+
+    const juliaPP = new JuliaSetPostProcess(julia, scene, []);
+    scene.cameras.forEach(camera => camera.attachPostProcess(juliaPP));
+    scene.onNewCameraAddedObservable.add((camera) => {
+        camera.attachPostProcess(juliaPP);
+    });
+
+    scene.onBeforeRenderObservable.add(() => {
+        const deltaSeconds = engine.getDeltaTime() / 1000;
+        juliaPP.update(deltaSeconds);
+    });
+
+    return julia.getTransform();
+}
+
+let targetObject: TransformNode;
+
+const sceneType = urlParams.get("scene");
+
+if (sceneType === "mandelbulb") {
+    targetObject = createMandelbulb();
+} else if (sceneType === "julia") {
+    targetObject = createJulia();
+} else {
+    targetObject = createMandelbulb();
+}
+
+const xr = await scene.createDefaultXRExperienceAsync();
+if(xr.baseExperience) {
+    // web xr code goes here
+    const xrCamera = xr.baseExperience.camera;
+    xrCamera.setTransformationFromNonVRCamera(camera, true);
+}
+
+scene.onBeforeCameraRenderObservable.add((camera) => {
+    depthRenderer.getDepthMap().activeCamera = camera;
+});
 
 scene.executeWhenReady(() => {
     engine.loadingScreen.hideLoadingUI();
